pyproject.toml
[project]
name = "ansible-collection-customapi-inventory"
version = "0.1.0"
description = "An Ansible collection providing a dynamic inventory plugin for a custom API."
authors = [
    {name = "Jeongseup",email = "seup87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansible-core (>=2.18.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [
    { include = "plugins" }, # 'plugins' 디렉토리를 패키지에 포함
]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.1"
poethepoet = "^0.34.0"

[tool.poe.tasks]
# 작업 이름 (예: test-inventory)을 자유롭게 지정할 수 있습니다.
# 명령어는 프로젝트 루트 디렉토리에서 실행되는 것을 기준으로 작성합니다.
[tool.poe.tasks.test-inventory] # 작업 이름은 test-inventory 또는 원하는 다른 이름
shell = "ANSIBLE_CONFIG=./tests/integration/ansible.cfg ansible-inventory --list"
help = "Runs ansible-inventory for integration testing (using shell execution)"

[tool.poe.tasks.test-vars] # 작업 이름은 test-vars 또는 원하는 다른 이름
shell = "ANSIBLE_CONFIG=./tests/integration/ansible.cfg ansible-inventory --vars --list"
help = "Runs ansible-inventory variables"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
